+++ rex/tasks/simple_task.py
@@ -411,9 +411,7 @@ class SimpleTask(TaskBase):
         eval_loader = self.get_data_loader(
             dataset_name, is_eval=True, epoch=self.history["curr_epoch"]
         )
-        loader = pbar(
-            eval_loader, desc=f"{dataset_name} - {postfix} Eval", ascii=True
-        )
+        loader = pbar(eval_loader, desc=f"{dataset_name} - {postfix} Eval", ascii=True)
 
         eval_loss = 0.0
         metrics = {}